linkedin/requirements.txt
fastapi
uvicorn
python-dotenv
groq
selenium
webdriver-manager
reportlab
pydantic
email-validator